Verify HTTP Signature headers
Implement verification of the draft spec HTTP Signature headers that
Drone uses to provide a level of webhook authenticity with a PSK. Drone
can be configured to generate the signature header by setting the
DRONE_WEBHOOK_SECRET= environment variable. In combination with the
digest authentication, drone-notify should have reasonable certainty
that the webhook events from Drone are authentic and the message body
has not been tampered with in-flight.

Closes: https://github.com/spritsail/drone-notify/issues/9
Replaces: https://github.com/spritsail/drone-notify/issues/10
Co-authored-by: Davide Perozzi <davide@undesigned.studio>
Co-authored-by: Joe Groocock <me@frebib.net>
Signed-off-by: Joe Groocock <me@frebib.net>